% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_SOL.R
\name{file_exists_on_server}
\alias{file_exists_on_server}
\title{Check if a File Exists on Server}
\usage{
file_exists_on_server(url)
}
\arguments{
\item{url}{Character. The URL of the file.}
}
\value{
Logical. `TRUE` if the file exists, `FALSE` otherwise.
}
\description{
Sends an HTTP HEAD request to check if a file is available at the specified URL.
}
