% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_to_SPC.R
\name{ras_to_SPC}
\alias{ras_to_SPC}
\title{Convert RasterStack to SoilProfileCollection}
\usage{
ras_to_SPC(rstack)
}
\arguments{
\item{rstack}{A \code{SpatRaster} with layer names indicating depth intervals (e.g., "claytotal_r_0_5")}
}
\value{
A \code{SoilProfileCollection} object
}
\description{
Convert RasterStack to SoilProfileCollection
}
