% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLUS.R
\name{fetch_SOL}
\alias{fetch_SOL}
\title{Fetch SOLUS100 Soil Data}
\usage{
fetch_SOL(aoi, properties, depths, measures, output_dir, suffix = "")
}
\arguments{
\item{aoi}{SpatRaster or SpatVector. Area of interest.}

\item{properties}{Character vector. List of soil properties to download.}

\item{depths}{Character vector. Soil depth intervals to download.}

\item{measures}{Character vector. Measures to download ("l", "h", "p", "rpi").}

\item{output_dir}{Character. Directory where files will be saved.}

\item{suffix}{Character. Optional suffix for filenames.}
}
\value{
Saves downloaded and clipped rasters.
}
\description{
Downloads and processes soil property data from SOLUS100.
}
\examples{
\dontrun{
# Define AOI and download SOLUS100 data
aoi <- rast("path_to_aoi.tif")
output_directory <- "path/to/your/directory"
fetch_SOL(
  aoi = aoi,
  properties = c("claytotal", "soc"),
  depths = c("0_cm", "30_cm"),
  measures = c("p", "l", "h"), #rpi is data heavy
  output_dir = "output_directory",
  suffix = "test"
)
}

}
