% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_to_SPC.R
\name{pts_to_SPC}
\alias{pts_to_SPC}
\title{Convert Extracted Point Soil Data to SoilProfileCollection}
\usage{
pts_to_SPC(rstack, locations, id_column = "Name")
}
\arguments{
\item{rstack}{A \code{SpatRaster} object}

\item{locations}{An \code{sf} or \code{SpatVector} object of point locations}

\item{id_column}{A column name in \code{locations} that uniquely identifies each point (e.g., "Name")}
}
\value{
A \code{SoilProfileCollection} object
}
\description{
Convert Extracted Point Soil Data to SoilProfileCollection
}
